Serpentine Heater design ready for try
both a 22 ga and a 26 ga jig, stls. Both jigs are a little short,
because the last arc and turn closest to the apex doesn't fit. It should
be a small matter. Also the 22 ga design is a bit short (gap should be
less then 8, but its 10) and may require a slightly lower voltage